requirements.txt
# Core Dependencies
python-dotenv>=1.0.0

# PDF Processing
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
Pillow>=10.0.0

# AI/ML APIs
openai>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.25.0

# Optional: Google Cloud (for Vertex AI)
google-cloud-aiplatform>=1.38.0

# Utilities
requests>=2.31.0
